requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
redis==5.0.1
celery==5.3.6
stripe==7.11.0
Pillow==10.2.0
django-filter==23.5
django-storages==1.14.2
boto3==1.34.34
python-dotenv==1.0.1
django-rest-knox==4.2.0 